% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\name{msg}
\alias{msg}
\title{: msg: Formats and prints a message to screen.}
\usage{
msg(...)
}
\arguments{
\item{...}{A series of objects (usually strings and numbers) to concatenate
and print to screen.}
}
\description{
Message function
}
