% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\name{print.goalp}
\alias{print.goalp}
\title{: print.goalp: Prints a summary of a goalp object to screen.}
\usage{
\method{print}{goalp}(x, ...)
}
\arguments{
\item{x}{A goalp object.}

\item{...}{Additional arguments. Ignored.}
}
\value{
A scalar character (i.e. a text string) with a human-readable
        formulation of the goal programming problem represented by
        goalp object \code{x}. This can be edited and used as an input
        to \link{goalp}, if modifications to the goal programming problem
        are required.
}
\description{
Prints a human-readable formulation of a goal programming problem.
}
