% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\name{summary.goalp}
\alias{summary.goalp}
\title{: summary.goalp: Prints a summary of a goalp object to screen.}
\usage{
\method{summary}{goalp}(object, ...)
}
\arguments{
\item{object}{A goalp object.}

\item{...}{Additional arguments. Ignored.}
}
\value{
No return value (NULL). Called for its side effect of
        printing a summary of a goalp object to the standard output
        (usually the console).
}
\description{
Prints a summary of a goalp object to the console.
}
