% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\name{validate_goalp}
\alias{validate_goalp}
\title{: validate_goalp: A validator for goalp objects.}
\usage{
validate_goalp(gp)
}
\arguments{
\item{gp}{A goalp object.}
}
\value{
The unmodified input invisibly.
}
\description{
Checks that the internals of a goalp object are consistent.
}
